Renamed to fit now that the class is now utility not service
